Create README.md
International investment bank was struggling to achieve more stable development practices within the organization. Because banking environment is highly controlled and rigid along with other complexities, it’s difficult to makechanges which are quick and agile.They brought git in their environment and understood how transforming it was for just everything. They explored the new agile methods and do away with old traditional development styles and reinvented their stack to